requirements.txt
accelerate==0.19.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==23.1.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.3
docker-pycreds==0.4.0
filelock==3.12.0
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.15.1
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
joblib==1.2.0
jsonlines==3.1.0
nltk==3.8.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
packaging==23.1
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.23.2
psutil==5.9.5
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.25.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.6.3
urllib3==2.0.2
wandb==0.15.4
zipp==3.15.0
